save ext request dev status

+++ roles/ui/files/FWO.UI/Services/WfDbAccess.cs
@@ -73,7 +73,7 @@ namespace FWO.Ui.Services
             return ticketsOut;
         }
 
-        public async Task<WfTicket> GetTicket(int id)
+        public async Task<WfTicket> GetTicket(long id)
         {
             WfTicket ticket = new ();
             try
